properties massless in SingleStar class instead
